+++ web/src/services/mod.rs
@@ -1,4 +1,5 @@
 pub mod api;
+pub mod authentication_token_service;
 pub mod integration_connection_service;
 pub mod nango;
 pub mod notification_service;